15_A2A_LangGraph/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0

# HTTP client for A2A protocol communication
httpx>=0.25.0

# Async support
asyncio

# Environment and configuration
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.0.0

# Logging and utilities
pydantic>=2.0.0 